Moves determining content types into theme_client
This ensures that users of phoenix.LoadAsset will have the data
correctly loaded based on content type. Binary data will be set as
the attachment, compared to the value which was done previously.

Though, theme-watch pretty much has the same implementation of LoadAsset
which should be resolved in order to reduce unnecessary duplication.

fixes #11