Does a bit more refactoring to ThemeClient
There was a bunch of stuff going on in the theme client code
that was really tough to read through and look at. This moves
much of the response parsing into an APIThemeEvent object that
is responsible for pulling all the useful information out of
the API responses from Shopify and turning them into something
that can easily be worked with.

Also adds a super basic Theme struct to make decoding JSON suck
less