Updated Bootstrap, Configure, Download and Watch
Calls to these functions can now be made with various Options structs
that can be used instead of super annoying and long method signatures.

There is a caveat with the BasicOptions in that you cannot easily create
a substructure. This is a little weird/confusing and I'll need to do
some research on how to get around this. Might just have thing that
lets you dupe the data into the other object easily or something. Don't
quite know yet :/

Uses a global event log from the command line utility. Will need to
be sure that we don't allow it to ever get closed, but for now it'll
be fine.