Ignores files that are considered to be compiled
This is a pretty rough estimation as to which files are
compiled, but most of them typically follow a similar scheme
like "filename.js.liquid" and so on.

There is a chance this might cause some other kinds of bugs though,
so perhaps this isn't the right approach. Another alternative is
to simply detect those files and add them to the ignore section
of the environments configuration.