Initial work on checksum comparison
During deploy, watch and download, it will compare the checksums of the files on
shopify compared with local checksums and will skip files that are the same. It will
also pass the previous checksum in the requests to shopify in the
`checksum_lock` param to ensure the version being updated has not
changed since last observed. Lastly for json files, we compact them
before checksuming since that is how checksums are calculated on
shopify.